requirements.txt
torch==1.2.0
torchvision==0.4.0
numpy==1.19.1
Pillow==8.2.0
tensorboardX==2.2
opencv-python-headless==4.1.0.25
plyfile==0.7.4
open3d==0.8.0.0